+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added `-k/--insecure` flag to disable SSL certificate verification
 - Added `json` output feature to analyze module
+- Added `analyze` module auto-run to `run` module
 
 ### Changed